playing around

+++ app/src/main/java/com/example/hwada/adapter/FavoritesAdapter.java
@@ -28,7 +28,7 @@ public class FavoritesAdapter extends RecyclerView.Adapter<FavoritesAdapter.Favo
     public void onBindViewHolder(@NonNull FavoritesViewHolder holder, int position) {
         holder.jop.setText(list.get(position).getTitle());
         holder.fullName.setText(list.get(position).getAuthorName());
-        holder.distance.setText(list.get(position).getDistance());
+        holder.distance.setText((int) list.get(position).getDistance());
         holder.date.setText(list.get(position).getDate());
         holder.favImage.setImageResource(R.drawable.fav_checked_icon);
     }
